+++ src/test/java/io/github/astrapi69/crypto/pw/PasswordFileDecryptorTest.java
@@ -1,0 +1,109 @@
+/**
+ * The MIT License
+ *
+ * Copyright (C) 2015 Asterios Raptis
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package io.github.astrapi69.crypto.pw;
+
+import de.alpharogroup.file.delete.DeleteFileExtensions;
+import de.alpharogroup.file.search.PathFinder;
+import io.github.astrapi69.AbstractTestCase;
+import io.github.astrapi69.checksum.FileChecksumExtensions;
+import io.github.astrapi69.crypto.algorithm.MdAlgorithm;
+import io.github.astrapi69.crypto.algorithm.SunJCEAlgorithm;
+import io.github.astrapi69.crypto.file.FileDecryptor;
+import io.github.astrapi69.crypto.file.FileEncryptor;
+import io.github.astrapi69.crypto.model.CryptModel;
+import io.github.astrapi69.crypto.model.StringDecorator;
+import org.testng.annotations.BeforeMethod;
+import org.testng.annotations.Test;
+
+import javax.crypto.Cipher;
+import java.io.File;
+
+import static org.testng.Assert.*;
+import static org.testng.AssertJUnit.assertEquals;
+
+/**
+ * The unit test class for the class {@link PasswordFileDecryptor}
+ *
+ * @author Asterios Raptis
+ * @version 1.0
+ */
+public class PasswordFileDecryptorTest
+	extends AbstractTestCase<String, String>
+{
+
+	File cryptDir;
+	File decrypted;
+	PasswordFileDecryptor decryptor;
+	File encrypted;
+	PasswordFileEncryptor encryptor;
+	File toEncrypt;
+	String password;
+
+	/**
+	 * Sets up method will be invoked before every unit test method in this class
+	 */
+	@Override
+	@BeforeMethod
+	protected void setUp()
+	{
+		password = "foo";
+		cryptDir = new File(PathFinder.getSrcTestResourcesDir(), "crypt");
+		toEncrypt = new File(cryptDir, "test.txt");
+	}
+
+	/**
+	 * Test method for test the method {@link PasswordFileDecryptor#decrypt(File)}
+	 *
+	 * @throws Exception
+	 *             is thrown if any error occurs on the execution
+	 */
+	@Test public void testDecrypt() throws Exception
+	{
+
+		File encryptedCnstr;
+		File decryptedCnstr;
+		String encryptedFilename;
+		String decryptedFilename;
+		// new scenario...
+		encryptedFilename = "encryptedCnstr.enc";
+		decryptedFilename = "decryptedCnstr.decrypted";
+		encryptedCnstr = new File(cryptDir, encryptedFilename);
+		decryptedCnstr = new File(cryptDir, decryptedFilename);
+		encryptor = new PasswordFileEncryptor(password, encryptedCnstr);
+		encrypted = encryptor.encrypt(toEncrypt);
+
+		decryptor = new PasswordFileDecryptor(password, decryptedCnstr);
+
+		decrypted = decryptor.decrypt(encrypted);
+
+		expected = FileChecksumExtensions.getChecksum(toEncrypt, MdAlgorithm.MD5.name());
+		actual = FileChecksumExtensions.getChecksum(decrypted, MdAlgorithm.MD5.name());
+		assertEquals(actual, expected);
+		// clean up...
+		DeleteFileExtensions.delete(encrypted);
+		DeleteFileExtensions.delete(decrypted);
+	}
+
+}